// types.ts
export interface Investment {
  id: string;
  asset: string;
  type: 'Ações' | 'FIIs' | 'ETFs' | 'Cripto' | 'Renda Fixa';
  quantity: number;
  avgPrice: number;
  currentPrice: number;
  totalValue: number;
  profitability: number;
}